% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_make.R
\name{get_moron_pal}
\alias{get_moron_pal}
\title{Get a palette made by a Morin lab member}
\usage{
get_moron_pal(palette_name, num_col)
}
\arguments{
\item{num_col}{}
}
\value{

}
\description{
Get a palette made by a Morin lab member
}
