% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_make.R
\name{image_to_palette}
\alias{image_to_palette}
\title{Extract a palette from a colourful image
This can be slow to run so the expectation is that any useful palettes will be saved in this package}
\usage{
image_to_palette(image_file, num_col = 9)
}
\arguments{
\item{num_col}{}
}
\value{

}
\description{
Extract a palette from a colourful image
This can be slow to run so the expectation is that any useful palettes will be saved in this package
}
\examples{
pepper_image = system.file("extdata", "SgtPepper.jpg", package="moroncolours")
sgt_pepper_9 = image_to_palette(pepper_image,9)
hne_image = system.file("extdata", "HandE.jpg", package="moroncolours")
hne_9 = image_to_palette(hne_image,9)
led_zep_image = system.file("extdata","ledzep.jpg",package="moroncolours")
led_zep_9 = image_to_palette(led_zep_image,9)
}
