% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_make.R
\name{save_moron_pal}
\alias{save_moron_pal}
\title{Save a new palette for future use.
Running this will create a new value in the moron_db object
If you want to ensure the palette can be recreated add the call
to this function to palette_generato.R}
\usage{
save_moron_pal(palette_colours, palette_name, num_col)
}
\arguments{
\item{num_col}{}
}
\value{

}
\description{
Save a new palette for future use.
Running this will create a new value in the moron_db object
If you want to ensure the palette can be recreated add the call
to this function to palette_generato.R
}
\examples{
save_moron_pal(led_zep_9,"ledzep",9)
}
